+++ src/pages/api/transactions/[id].ts
@@ -2,7 +2,13 @@ import { PrismaClient, Transaction } from "@prisma/client";
 import { withIronSessionApiRoute } from "iron-session/next";
 import { NextApiHandler } from "next";
 import { sessionSettings } from "../../../sessions/ironSessionSettings";
-import { getOptionalValue, getValue, requireAuthentication } from "../../../utils/apiUtils";
+import {
+  getOptionalValue,
+  getValue,
+  requireAuthentication,
+  requireResourceAccess,
+  StatusCodes
+} from "../../../utils/apiUtils";
 import { numberValidator, stringValidator, uuidValidator } from "../../../utils/apiValidators";
 import { withApiErrorHandling } from "../../../utils/errorHandling";
 
@@ -10,13 +16,14 @@ const handler: NextApiHandler = async (req, res) => {
   if (req.method === "DELETE") {
     requireAuthentication(req, "You must be logged in to delete a transaction");
 
-    // TODO: Check that the user has access to the transaction
-    const id = getValue(req.query, "id", uuidValidator);
-
+    const transactionId = getValue(req.query, "id", uuidValidator);
     const prisma = new PrismaClient();
+
+    await requireResourceAccess(req.session.user.id, transactionId, "transaction", prisma);
+
     await prisma.transaction.delete({
       where: {
-        id: String(id)
+        id: transactionId
       }
     });
 
@@ -25,9 +32,7 @@ const handler: NextApiHandler = async (req, res) => {
   if (req.method === "PUT") {
     requireAuthentication(req, "You must be logged in to update a transaction");
 
-    // TODO: Check that the user has access to the sink, storage and transaction
-
-    const id = getValue(req.query, "id", uuidValidator);
+    const transactionId = getValue(req.query, "id", uuidValidator);
 
     const editedTransaction: Partial<Transaction> = {};
     editedTransaction.amount = getOptionalValue(req.body, "amount", numberValidator);
@@ -36,9 +41,18 @@ const handler: NextApiHandler = async (req, res) => {
     editedTransaction.storageId = getOptionalValue(req.body, "storageId", uuidValidator);
 
     const prisma = new PrismaClient();
+
+    await requireResourceAccess(req.session.user.id, transactionId, "transaction", prisma);
+    if (editedTransaction.sinkId) {
+      await requireResourceAccess(req.session.user.id, editedTransaction.sinkId, "sink", prisma);
+    }
+    if (editedTransaction.storageId) {
+      await requireResourceAccess(req.session.user.id, editedTransaction.storageId, "storage", prisma);
+    }
+
     const newTransaction = await prisma.transaction.update({
       where: {
-        id
+        id: transactionId
       },
       data: editedTransaction
     });
@@ -46,7 +60,7 @@ const handler: NextApiHandler = async (req, res) => {
     return res.json(newTransaction);
   }
   else {
-    res.status(405).send(null);
+    res.status(StatusCodes.MethodNotAllowed).send(null);
   }
 };
